# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SDIT-Group API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/notif": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get list of notification",
                "description": "Menampilkan data list notifikasi sesuai dengan email user",
                "operationId": "indexNotif",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/view-notif/{id}": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get detail of a notif",
                "description": "Menampilkan detail notifikasi berdasarkan ID",
                "operationId": "showNotif",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari notif yang ingin diambil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Boim has Assign to you : Lorem EDIT"
                                                },
                                                "action": {
                                                    "type": "string",
                                                    "example": "Asign"
                                                },
                                                "link_url": {
                                                    "type": "string",
                                                    "example": "view-ticket/1"
                                                },
                                                "status_read": {
                                                    "type": "string",
                                                    "example": "Not Read"
                                                },
                                                "read_on": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "31-01-2025"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}